[scroll_overlay] Measure velocity at each frame
Even when measuring based on the real time, if we take measurements
on a periodic timer then the results show an occasional flicker
because the new offset is recomputed once per frame, and so
occasionally an interval sees one more or fewer updates than usual
because it just barely covers a new frame at both ends or just barely
misses one at each end.

Instead, check in just after each frame.

This pretty much eliminates flicker in the display of the Flutter
velocity.