[scroll_overlay] Add debug print on createBallisticSimulation
This has been extremely convenient in my recent scrolling
investigations.

Because the ScrollPhysics protocol is a bit subtle to hook into
correctly, and requires some boilerplate, inserting debug logging
for it is a bit more complicated than just writing down a
`debugPrint` call of one's choice.  So it benefits from having
this permanent infrastructure in the app.