[scroll_overlay] Upgrade Android Gradle Plugin to 4.x; Gradle to 6.7.x
I've started seeing the symptoms of flutter/flutter#53657 when I
try to run this app.  That blocks building and running it.

I don't totally understand why that's happening, as that issue is
said to only appear on a range of AGP versions later than the 3.5.0
this is using.  Nor do I understand why it started happening for me
when it wasn't before.

But regardless of the cause, thankfully the symptoms are resolved by
following the recommendation in that issue thread:
  https://github.com/flutter/flutter/issues/53657#issuecomment-795808016
to upgrade AGP.  And upgrading seems for the best anyway.

We go for the latest 4.x version, as called for there.  I just let
Android Studio do the upgrade according to its recommendations, and
then bumped compileSdkVersion to the new requirement.